+++ src/index.js
@@ -5,14 +5,19 @@ import { Provider } from 'react-redux';
 import { Router, Route, browserHistory } from 'react-router';
 import { syncHistoryWithStore } from 'react-router-redux';
 import Home from './home';
-import store from './store';
+import configureStore from './store';
+import DevTools from './dev-tools';
 
+const store = configureStore();
 const history = syncHistoryWithStore(browserHistory, store);
 
 ReactDOM.render(
   <Provider store={store}>
-    <Router history={history}>
-      <Route path="/*" component={Home} />
-    </Router>
+    <div>
+      <Router history={history}>
+        <Route path="/*" component={Home} />
+      </Router>
+      <DevTools />
+    </div>
   </Provider>,
 document.getElementById('container'));