+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ### 2016-03-14
 
 + First version
+
+### 2016-04-12
+
++ Add `redux-devtools`
++ Demo available: [treemap-with-router](http://kuy.github.io/treemap-with-router)